pyproject.toml

[project]
name = "underground"
version = "0.1.0"
description = ""
authors = [
    {name = "radomyr", email = "h.radomyr@proton.me"},
]
dependencies = [
    "pandas>=2.0.1",
    "openai>=0.27.6",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
]
